refactor: 💡 Refactored common types
